fix(cache): split llm_cache tests + fix invalidate_chat user_specific
llm_cache.rs was 839 lines (over the U-16 800-line ceiling) which
blocked any further edits to fix the user_specific invalidation bug.

Move the inline test module to llm_cache_tests.rs (327 lines) via
`#[cfg(test)] #[path = "llm_cache_tests.rs"] mod tests;`, bringing
the main file to 530 lines and unblocking the fix.

Then fix the actual bug:

invalidate_chat ignored config.user_specific and built the cache key
without a user id, even though writes (cache_chat_response_with_user)
correctly key by user when user_specific is enabled. The result:
calling invalidate_chat deleted a key that was never written, leaving
the per-user entry live until TTL — stale responses leaked across
users.

Add invalidate_chat_with_user(&self, request, user_id) that mirrors
the cache_chat_response_with_user write-side branch. Have the bare
invalidate_chat delegate to it with user_id=None to preserve the
existing API for callers that don't have user-specific caching.

Add a regression test asserting the invalidate-with-user path
correctly removes the per-user entry.
